Remove the Gzip/Gunzip code
For compression, we're probably better off using .parquet files.
This makes the tar plugin simpler as we no longer have to maintain a read cache
We keep the .gztar URI scheme just in case someelse creates a .tar NIO2 plugin
