+++ src/resources/META-INF/services/java.nio.file.spi.FileSystemProvider
@@ -1,1 +1,1 @@
-org.epics.archiverappliance.utils.nio.gztar.GZTarFileSystemProvider
+org.epics.archiverappliance.utils.nio.tar.TarFileSystemProvider
