+++ src/test/org/epics/archiverappliance/utils/nio/tar/ConversionTest.java
@@ -1,4 +1,4 @@
-package org.epics.archiverappliance.utils.nio.gztar;
+package org.epics.archiverappliance.utils.nio.tar;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.logging.log4j.LogManager;
@@ -16,6 +16,7 @@ import org.epics.archiverappliance.config.ArchDBRTypes;
 import org.epics.archiverappliance.config.ConfigServiceForTests;
 import org.epics.archiverappliance.config.StoragePluginURLParser;
 import org.epics.archiverappliance.data.ScalarValue;
+import org.epics.archiverappliance.etl.conversion.ThruNumberAndStringConversion;
 import org.epics.archiverappliance.retrieval.postprocessors.DefaultRawPostProcessor;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -25,30 +26,33 @@ import org.junit.jupiter.api.Test;
 import java.io.File;
 import java.io.IOException;
 import java.net.URLEncoder;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
+import java.util.Collections;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Callable;
 
 /*
- * Test the storage plugin's rename function
+ * Test the storage plugin's convert function
  */
-public class RenameTest {
+public class ConversionTest {
     private static final Logger logger = LogManager.getLogger();
-    private static final String rootFolderStr = ConfigServiceForTests.getDefaultPBTestFolder() + "/gztar/RenameTest";
-    private static final String pvNameOld = "epics:arch:gztartest";
-    private static final String chunkKeyOld = pvNameOld.replace(":", File.separator);
-    private static final String pvNameNew = "epics:arch:gztartest:renamed";
-    private static final String chunkKeyNew = pvNameNew.replace(":", File.separator);
+    private static final String rootFolderStr =
+            ConfigServiceForTests.getDefaultPBTestFolder() + "/gztar/ConversionTest";
+    private static final String pvName = "epics:arch:gztartest";
+    private static final String chunkKey = pvName.replace(":", File.separator);
     private static ConfigServiceForTests configService;
 
     @BeforeAll
     public static void setUp() throws Exception {
         File rootFolder = new File(rootFolderStr);
         FileUtils.deleteDirectory(rootFolder);
-        Path pvPath = Paths.get(rootFolderStr, chunkKeyOld);
+        Path pvPath = Paths.get(rootFolderStr, chunkKey);
         logger.debug("Creating folder {}", pvPath.getParent().toFile().toString());
         assert pvPath.getParent().toFile().mkdirs();
         configService = new ConfigServiceForTests(1);
@@ -69,7 +73,7 @@ public class RenameTest {
             short year = forYear;
             for (int day = 0; day < 365; day++) {
                 ArrayListEventStream testData = new ArrayListEventStream(
-                        24 * 60 * 60, new RemotableEventStreamDesc(ArchDBRTypes.DBR_SCALAR_DOUBLE, pvNameOld, year));
+                        24 * 60 * 60, new RemotableEventStreamDesc(ArchDBRTypes.DBR_SCALAR_DOUBLE, pvName, year));
                 int startofdayinseconds = day * 24 * 60 * 60;
                 for (int secondintoday = 0; secondintoday < 24 * 60 * 60; secondintoday += skipSeconds) {
                     Instant dataTs = TimeUtils.convertFromYearSecondTimestamp(
@@ -82,18 +86,19 @@ public class RenameTest {
                                     0)
                             .makeClone());
                 }
-                storagePlugin.appendData(context, pvNameOld, testData);
+                storagePlugin.appendData(context, pvName, testData);
             }
         }
 
-        EAATar tarFile = new EAATar(rootFolderStr + File.separator + chunkKeyOld + ".tar");
+        EAATar tarFile = new EAATar(rootFolderStr + File.separator + chunkKey + ".tar");
         Map<String, TarEntry> entries = tarFile.loadCatalog();
         Assertions.assertTrue(
                 entries.size() == expectedCatalogEntryCount,
                 "Expecting " + expectedCatalogEntryCount + " entries; got " + entries.size() + " entries");
     }
 
-    private void testRetrieval(String pvName, Instant start, Instant end, int expectedEventCount) throws Exception {
+    private void testRetrieval(Instant start, Instant end, int expectedEventCount, ArchDBRTypes expectedType)
+            throws Exception {
         StoragePlugin storagePlugin = StoragePluginURLParser.parseStoragePlugin(
                 "pb://localhost?name=Test&rootFolder=" + URLEncoder.encode("gztar://" + rootFolderStr, "UTF-8")
                         + "&partitionGranularity=PARTITION_DAY",
@@ -107,13 +112,15 @@ public class RenameTest {
             for (Callable<EventStream> callable :
                     storagePlugin.getDataForPV(context, pvName, start, end, new DefaultRawPostProcessor())) {
                 try (EventStream strm = callable.call()) {
+                    Assertions.assertEquals(
+                            strm.getDescription().getArchDBRType(),
+                            expectedType,
+                            "Expected " + expectedType + " got "
+                                    + strm.getDescription().getArchDBRType());
                     for (Event ev : strm) {
                         // logger.debug("Found event at {} {}",
                         // TimeUtils.convertToISO8601String(ev.getEventTimeStamp()),
                         // TimeUtils.convertToHumanReadableString(ev.getEventTimeStamp()));
-                        Assertions.assertEquals(
-                                ev.getEventTimeStamp().getEpochSecond(),
-                                ev.getSampleValue().getValue().doubleValue());
                         eventCount++;
                     }
                 }
@@ -125,13 +132,14 @@ public class RenameTest {
         }
     }
 
-    private void renamePV() throws IOException {
+    private void convertPV() throws IOException {
+        logger.debug("Starting conversion to DBR_SCALAR_FLOAT");
         StoragePlugin storagePlugin = StoragePluginURLParser.parseStoragePlugin(
                 "pb://localhost?name=Test&rootFolder=" + URLEncoder.encode("gztar://" + rootFolderStr, "UTF-8")
                         + "&partitionGranularity=PARTITION_DAY",
                 configService);
         try (BasicContext context = new BasicContext()) {
-            storagePlugin.renamePV(context, pvNameOld, pvNameNew);
+            storagePlugin.convert(context, pvName, new ThruNumberAndStringConversion(ArchDBRTypes.DBR_SCALAR_FLOAT));
         }
     }
 
@@ -140,40 +148,38 @@ public class RenameTest {
         short currentYear = TimeUtils.getCurrentYear();
         appendAndTestForYear(currentYear, 365, 60);
 
-        // Test retrieval before rename
+        // Test retrieval before convert
         Instant end = TimeUtils.getStartOfYear(currentYear).plusSeconds(24 * 60 * 60 * 32);
         for (int days = 1; days < 5; days++) {
             Instant start = end.minus(days, ChronoUnit.DAYS);
             long before = System.currentTimeMillis();
-            testRetrieval(pvNameOld, start, end, (days * 24 * 60) + 1);
+            testRetrieval(start, end, (days * 24 * 60) + 1, ArchDBRTypes.DBR_SCALAR_DOUBLE);
             long after = System.currentTimeMillis();
             logger.debug("Took {}(ms) to retrieve data for {} days", (after - before), days);
         }
 
-        for (int days = 1; days < 5; days++) {
-            Instant start = end.minus(days, ChronoUnit.DAYS);
-            long before = System.currentTimeMillis();
-            testRetrieval(pvNameNew, start, end, 0);
-            long after = System.currentTimeMillis();
-            logger.debug("Took {}(ms) to retrieve data for {} days", (after - before), days);
+        convertPV();
+
+        Path tarPath = Paths.get(rootFolderStr, chunkKey + ".tar");
+        Assertions.assertTrue(Files.exists(tarPath), "Expecting tar file to be present after conversion");
+        EAATar tarFile = new EAATar(tarPath.toString());
+        logger.debug("Loading catalog for {}", tarFile.getTarFileName());
+        Map<String, TarEntry> entries = tarFile.loadCatalog();
+        logger.debug("After conversion, we have {} entries in the catalog", entries.size());
+        List<String> entryNames = new LinkedList<String>(entries.keySet());
+        Collections.sort(entryNames);
+        // Assertions.assertTrue(
+        //         entryNames.size() == 4 * 365, "Expecting " + 4 * 365 + " entries; got " + entries.size() + "
+        // entries");
+        for (String entryName : entryNames) {
+            logger.debug("Entry {}", entryName);
         }
 
-        renamePV();
-
-        // Test retrieval before rename
-        // Rename does not delete data under the older name
+        // Test retrieval after convert
         for (int days = 1; days < 5; days++) {
             Instant start = end.minus(days, ChronoUnit.DAYS);
             long before = System.currentTimeMillis();
-            testRetrieval(pvNameOld, start, end, (days * 24 * 60) + 1);
-            long after = System.currentTimeMillis();
-            logger.debug("Took {}(ms) to retrieve data for {} days", (after - before), days);
-        }
-
-        for (int days = 1; days < 5; days++) {
-            Instant start = end.minus(days, ChronoUnit.DAYS);
-            long before = System.currentTimeMillis();
-            testRetrieval(pvNameNew, start, end, (days * 24 * 60) + 1);
+            testRetrieval(start, end, (days * 24 * 60) + 1, ArchDBRTypes.DBR_SCALAR_FLOAT);
             long after = System.currentTimeMillis();
             logger.debug("Took {}(ms) to retrieve data for {} days", (after - before), days);
         }
