+++ src/test/org/epics/archiverappliance/utils/nio/tar/GetDataAtTimeTest.java
@@ -5,7 +5,7 @@
  * EPICS archiver appliance is distributed subject to a Software License Agreement found
  * in file LICENSE that is included with this distribution.
  *******************************************************************************/
-package org.epics.archiverappliance.utils.nio.gztar;
+package org.epics.archiverappliance.utils.nio.tar;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.logging.log4j.LogManager;
