+++ src/test/org/epics/archiverappliance/utils/nio/tar/PBPluginTest.java
@@ -1,5 +1,7 @@
-package org.epics.archiverappliance.utils.nio.gztar;
+package org.epics.archiverappliance.utils.nio.tar;
 
+import edu.stanford.slac.archiverappliance.plain.PathNameUtility;
+import edu.stanford.slac.archiverappliance.plain.PathResolver;
 import org.apache.commons.io.FileUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -16,15 +18,14 @@ import org.epics.archiverappliance.config.ArchDBRTypes;
 import org.epics.archiverappliance.config.ConfigServiceForTests;
 import org.epics.archiverappliance.config.StoragePluginURLParser;
 import org.epics.archiverappliance.data.ScalarValue;
-import org.epics.archiverappliance.etl.conversion.ThruNumberAndStringConversion;
 import org.epics.archiverappliance.retrieval.postprocessors.DefaultRawPostProcessor;
+import org.epics.archiverappliance.utils.nio.ArchPaths;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
-import java.io.IOException;
 import java.net.URLEncoder;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -38,12 +39,11 @@ import java.util.Map;
 import java.util.concurrent.Callable;
 
 /*
- * Test the storage plugin's convert function
+ * Create gztar files thru the PlainPBStoragePlugin.
  */
-public class ConversionTest {
+public class PBPluginTest {
     private static final Logger logger = LogManager.getLogger();
-    private static final String rootFolderStr =
-            ConfigServiceForTests.getDefaultPBTestFolder() + "/gztar/ConversionTest";
+    private static final String rootFolderStr = ConfigServiceForTests.getDefaultPBTestFolder() + "/gztar/PBPluginTest";
     private static final String pvName = "epics:arch:gztartest";
     private static final String chunkKey = pvName.replace(":", File.separator);
     private static ConfigServiceForTests configService;
@@ -66,7 +66,8 @@ public class ConversionTest {
 
     private void appendAndTestForYear(short forYear, int expectedCatalogEntryCount, int skipSeconds) throws Exception {
         StoragePlugin storagePlugin = StoragePluginURLParser.parseStoragePlugin(
-                "pb://localhost?name=Test&rootFolder=" + URLEncoder.encode("gztar://" + rootFolderStr, "UTF-8")
+                "pb://localhost?name=Test&rootFolder="
+                        + URLEncoder.encode(URIUtils.GZTAR_SCHEME + "://" + rootFolderStr, "UTF-8")
                         + "&partitionGranularity=PARTITION_DAY",
                 configService);
         try (BasicContext context = new BasicContext()) {
@@ -91,16 +92,17 @@ public class ConversionTest {
         }
 
         EAATar tarFile = new EAATar(rootFolderStr + File.separator + chunkKey + ".tar");
+        logger.debug("Loading catalog for {}", tarFile);
         Map<String, TarEntry> entries = tarFile.loadCatalog();
         Assertions.assertTrue(
                 entries.size() == expectedCatalogEntryCount,
                 "Expecting " + expectedCatalogEntryCount + " entries; got " + entries.size() + " entries");
     }
 
-    private void testRetrieval(Instant start, Instant end, int expectedEventCount, ArchDBRTypes expectedType)
-            throws Exception {
+    private void testRetrieval(Instant start, Instant end, int expectedEventCount) throws Exception {
         StoragePlugin storagePlugin = StoragePluginURLParser.parseStoragePlugin(
-                "pb://localhost?name=Test&rootFolder=" + URLEncoder.encode("gztar://" + rootFolderStr, "UTF-8")
+                "pb://localhost?name=Test&rootFolder="
+                        + URLEncoder.encode(URIUtils.GZTAR_SCHEME + "://" + rootFolderStr, "UTF-8")
                         + "&partitionGranularity=PARTITION_DAY",
                 configService);
         logger.debug(
@@ -112,76 +114,103 @@ public class ConversionTest {
             for (Callable<EventStream> callable :
                     storagePlugin.getDataForPV(context, pvName, start, end, new DefaultRawPostProcessor())) {
                 try (EventStream strm = callable.call()) {
-                    Assertions.assertEquals(
-                            strm.getDescription().getArchDBRType(),
-                            expectedType,
-                            "Expected " + expectedType + " got "
-                                    + strm.getDescription().getArchDBRType());
                     for (Event ev : strm) {
                         // logger.debug("Found event at {} {}",
                         // TimeUtils.convertToISO8601String(ev.getEventTimeStamp()),
                         // TimeUtils.convertToHumanReadableString(ev.getEventTimeStamp()));
+                        Assertions.assertEquals(
+                                ev.getEventTimeStamp().getEpochSecond(),
+                                ev.getSampleValue().getValue().doubleValue());
                         eventCount++;
                     }
                 }
             }
             logger.debug("Done counting events");
             Assertions.assertTrue(
-                    eventCount == expectedEventCount,
-                    "Expecting " + expectedEventCount + " events; got " + eventCount + " events");
-        }
-    }
-
-    private void convertPV() throws IOException {
-        logger.debug("Starting conversion to DBR_SCALAR_FLOAT");
-        StoragePlugin storagePlugin = StoragePluginURLParser.parseStoragePlugin(
-                "pb://localhost?name=Test&rootFolder=" + URLEncoder.encode("gztar://" + rootFolderStr, "UTF-8")
-                        + "&partitionGranularity=PARTITION_DAY",
-                configService);
-        try (BasicContext context = new BasicContext()) {
-            storagePlugin.convert(context, pvName, new ThruNumberAndStringConversion(ArchDBRTypes.DBR_SCALAR_FLOAT));
+                    Math.abs(eventCount - expectedEventCount) < 2,
+                    "Expecting " + expectedEventCount + " events; got "
+                            + eventCount + " events" + " when looking from "
+                            + TimeUtils.convertToISO8601String(start) + " to " + TimeUtils.convertToISO8601String(end));
         }
     }
 
     @Test
     public void testAppendThruPlugin() throws Exception {
         short currentYear = TimeUtils.getCurrentYear();
-        appendAndTestForYear(currentYear, 365, 60);
-
-        // Test retrieval before convert
-        Instant end = TimeUtils.getStartOfYear(currentYear).plusSeconds(24 * 60 * 60 * 32);
-        for (int days = 1; days < 5; days++) {
-            Instant start = end.minus(days, ChronoUnit.DAYS);
-            long before = System.currentTimeMillis();
-            testRetrieval(start, end, (days * 24 * 60) + 1, ArchDBRTypes.DBR_SCALAR_DOUBLE);
-            long after = System.currentTimeMillis();
-            logger.debug("Took {}(ms) to retrieve data for {} days", (after - before), days);
+        for (int y = 3; y >= 0; y--) {
+            appendAndTestForYear((short) (currentYear - y), 365 * (3 - y + 1), 60);
         }
 
-        convertPV();
-
-        Path tarPath = Paths.get(rootFolderStr, chunkKey + ".tar");
-        Assertions.assertTrue(Files.exists(tarPath), "Expecting tar file to be present after conversion");
-        EAATar tarFile = new EAATar(tarPath.toString());
-        logger.debug("Loading catalog for {}", tarFile.getTarFileName());
+        Files.walk(Paths.get(rootFolderStr))
+                .forEach(p -> logger.debug("---- " + p.toAbsolutePath().toString()));
+        EAATar tarFile = new EAATar(rootFolderStr + File.separator + chunkKey + ".tar");
+        logger.debug("Loading catalog for {}", tarFile);
         Map<String, TarEntry> entries = tarFile.loadCatalog();
-        logger.debug("After conversion, we have {} entries in the catalog", entries.size());
         List<String> entryNames = new LinkedList<String>(entries.keySet());
         Collections.sort(entryNames);
-        // Assertions.assertTrue(
-        //         entryNames.size() == 4 * 365, "Expecting " + 4 * 365 + " entries; got " + entries.size() + "
-        // entries");
-        for (String entryName : entryNames) {
-            logger.debug("Entry {}", entryName);
-        }
+        Assertions.assertTrue(
+                entryNames.size() == 4 * 365, "Expecting " + 4 * 365 + " entries; got " + entries.size() + " entries");
+        // for(String entryName : entryNames) {
+        //     logger.debug("Entry {}", entryName);
+        // }
 
-        // Test retrieval after convert
-        for (int days = 1; days < 5; days++) {
+        Files.walk(Paths.get(rootFolderStr))
+                .forEach(p -> logger.debug("---- " + p.toAbsolutePath().toString()));
+
+        // Test retrieval
+        Instant end = TimeUtils.getStartOfYear(currentYear - 1).plusSeconds(24 * 60 * 60 * (365 / 2));
+        for (int days = 1; days < 60; days++) {
             Instant start = end.minus(days, ChronoUnit.DAYS);
             long before = System.currentTimeMillis();
-            testRetrieval(start, end, (days * 24 * 60) + 1, ArchDBRTypes.DBR_SCALAR_FLOAT);
+            testRetrieval(
+                    start,
+                    end,
+                    (days * 24 * 60)
+                            + 1
+                            + 1); // One for the last known event and one for the event that's exactly at the end time.
             long after = System.currentTimeMillis();
             logger.debug("Took {}(ms) to retrieve data for {} days", (after - before), days);
+        }
+
+        // Before update
+        end = TimeUtils.getStartOfYear(currentYear).plusSeconds(24 * 60 * 60 * 32);
+        for (int days = 1; days < 10; days++) {
+            Instant start = end.minus(days, ChronoUnit.DAYS);
+            testRetrieval(start, end, (days * 24 * 60) + 1 + 1);
+        }
+
+        // Delete the files that we are updating ( otherwise we'll be appending and we'll get the wrong result )
+        logger.debug("Deleting some files now");
+        try (ArchPaths archPaths = new ArchPaths()) {
+            String currYearStr = Short.toString(currentYear);
+            for (Path path : PathNameUtility.getAllPathsForPV(
+                    archPaths,
+                    URIUtils.GZTAR_SCHEME + "://" + rootFolderStr,
+                    pvName,
+                    ".pb",
+                    PathResolver.BASE_PATH_RESOLVER,
+                    configService.getPVNameToKeyConverter())) {
+                if (path.toString().contains(currYearStr)) {
+                    logger.debug("Deleting {}", path.toString());
+                    Files.delete(path);
+                }
+            }
+        }
+        logger.debug("Done deleting some files now");
+
+        logger.info("Updating some PB files now");
+        appendAndTestForYear(currentYear, 365 * 4, 120);
+
+        // After update
+        end = TimeUtils.getStartOfYear(currentYear).plusSeconds(24 * 60 * 60 * 32);
+        for (int days = 1; days < 10; days++) {
+            Instant start = end.minus(days, ChronoUnit.DAYS);
+            testRetrieval(
+                    start,
+                    end,
+                    ((days * 24 * 60) / 2)
+                            + 1
+                            + 1); // We generate data once every two minutes; so we should halve the number of samples
         }
     }
 }
